% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_array_na.rm}
\alias{add_array_na.rm}
\title{Add two arrays or matrices, ignoring NA values by default}
\usage{
add_array_na.rm(ary1, ary2, na.rm = TRUE)
}
\arguments{
\item{ary1, ary2}{numeric arrays or matrices. The arrays to be added. They must be of the same dimension.}

\item{na.rm}{logical(1). TRUE (default) if missing values (\code{NA}) should be ignored in the summation. If both elements in a given position are missing, then the result will be  \code{NA}.}
}
\value{
An array or matrix of the same dimensions as \code{ary1} and \code{ary2} whose values are the sums of \code{ary1} and \code{ary2} in each corresponding element.
}
\description{
Array or matrix addition in base R sets all sums in an element position to NA if any element in that position is NA in either of the arrays being added. In contrast, this function ignores NA values by default in its addition.
}
\examples{
(x1 <- matrix(c(NA,NA,2,2),2,2))
(x2 <- matrix(c(NA,3,NA,NA),2,2))
(x3 <- matrix(c(NA,NA,NA,NA),2,2))
(x4 <- matrix(c(1,2,3,4),2,2))

add_array_na.rm(x1, x2)
add_array_na.rm(x1, x3)
add_array_na.rm(x1, x4)

Reduce(add_array_na.rm, list(x1, x2, x3))

}
