% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_stats.R
\name{ale_stats}
\alias{ale_stats}
\title{ale_stats.R}
\usage{
ale_stats(ale_y, ale_n, ecdf_pos_y, ecdf_neg_y, zeroed_ale = FALSE)
}
\arguments{
\item{ecdf_pos_y, ecdf_neg_y}{ecdf function. ECDF functions of the upper and lower
halves respectively of the Y values relative to the median. (The precise median is
arbitrarily placed in the lower half: neg is <= median; pos is > median).}
}
\description{
Calculate statistical measures based on ALE
}
