% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots.R
\name{ALEPlots}
\alias{ALEPlots}
\title{ALE plots with print and plot methods}
\usage{
ALEPlots(
  obj,
  ...,
  ale_centre = "median",
  consolid_cats = 10L,
  y_1d_refs = c("25\%", "75\%"),
  rug_sample_size = obj@params$sample_size,
  min_rug_per_interval = 1L,
  min_col_width = 0.05,
  y_nonsig_band = 0.05,
  seed = 0,
  silent = FALSE
)
}
\arguments{
\item{obj}{\code{ALE} or \code{ModelBoot} object. The object containing ALE data to be plotted.}

\item{...}{not used. Inserted to require explicit naming of subsequent arguments.}

\item{ale_centre}{character(1) in c('median', 'mean', 'zero'). The ALE y values in the plots will be centred relative to this value. 'median' is the default. 'zero' will maintain the actual ALE values, which are centred on zero.}

\item{consolid_cats}{integer(1) > 1 or list. For 1D plots of categorical variables, only a maximum of \code{consolid_cats} distinct values (e.g., factor levels) are shown. The top \code{consolid_cats - 1} values in ALE strength are shown and all other values are consolidated into an "other" category. See details for the calculation. Valid formats are:
\itemize{
\item Single integer > 1: Consolidate categories only if there are more than \code{consolid_cats} factor levels.
\item List with required levels: a list with exactly two elements: \code{max} is the same as the single-integer option above: the maximum allowable levels before consolidation begins; \code{include} is a named list. Each sublist is a character vector of specific levels that must be included for the factor variable that it names. Unknown names or factor levels trigger an error.
An example of the list format would be:
\code{consolid_cats = list(max = 10, include = list(model = c("Cadillac Fleetwood", "Volvo 142E")))}
}}

\item{y_1d_refs}{character or numeric vector. For 1D ALE plots, the y outcome values for which a reference line should be drawn. If a character vector, \code{y_1d_refs} values are names from \code{obj@params$y_summary} (usually quantile names). If a numeric vector, \code{y_1d_refs} values must be values within the range of y, that is, between \code{obj@params$y_summary$min} and \code{obj@params$y_summary$max} inclusive.}

\item{rug_sample_size, min_rug_per_interval}{non-negative integer(1). Rug plots are down-sampled to \code{rug_sample_size} rows, otherwise they can be very slow for large datasets. By default, their size is the value of \code{obj@params$sample_size}. They maintain representativeness of the data by guaranteeing that each of the ALE bins will retain at least \code{min_rug_per_interval} elements; usually set to just 1 (default) or 2. To prevent this down-sampling, set \code{rug_sample_size} to \code{Inf} (but then the \code{ALEPlots} object would store the entire dataset, so could become very large).}

\item{min_col_width}{numeric(1) in \code{[0.01, 1]}. Column charts scale each column such that the column representing the category with the most elements has a scale of 1 and all other columns have a width that is a fraction of the largest category proportional to their numbers of elements. However, for visibility, no column is displayed narrower than a scale of \code{min_col_width}. To disable scaling by width, set \code{min_col_width = 1}.}

\item{y_nonsig_band}{numeric(1) from 0 to 1. If there are no p-values, some plots (notably the 1D effects plot) will shade grey the inner \code{y_nonsig_band} quantile below and above the \code{ale_centre} average (the median, by default) to indicate nonsignificant effects.}

\item{seed}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{silent}{See documentation for \code{\link[=ALE]{ALE()}}}
}
\value{
An object of class \code{ALEPlots} with properties \code{plots} and \code{params}.
}
\description{
An \code{ALEPlots} S7 object contains the ALE plots from \code{ALE} or \code{ModelBoot} objects stored as \code{ggplot} objects. The \code{ALEPlots} constructor creates all possible plots from the \code{ALE} or \code{ModelBoot} passed to it---not only individual 1D and 2D ALE plots, but also special plots like the ALE effects plot. So, an \code{ALEPlots} object is a collection of plots, almost never a single plot. To retrieve specific plots, use the \link[=get-ALEPlots-method]{get() method}. See the examples with the \code{\link[=ALE]{ALE()}} and \code{\link[=ModelBoot]{ModelBoot()}} objects for how to manipulate \code{ALEPlots} objects.
}
\section{Methods}{

\code{ALEPlots} objects support \link[=get-ALEPlots-method]{get()},
\link[=plot-ALEPlots-method]{plot()}, \link[=print-ALEPlots-method]{print()},
\link[=subset-ALEPlots-method]{subset()}, and
\link[=summary-ALEPlots-method]{summary()}. Each method topic documents its
complete signature and class-specific arguments.
}

\section{Properties}{

\describe{
\item{plots}{Stores the ALE plots. Use the \link[=get-ALEPlots-method]{get() method} to access them.}
\item{params}{The parameters used to calculate the ALE plots. These include most of the arguments used to construct the \code{ALEPlots} object. These are either the values provided by the user or used by default if the user did not change them but also includes several objects that are created within the constructor. These extra objects are described here, as well as those parameters that are stored differently from the form in the arguments:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* `y_col`, `y_cats`: See documentation for [ALE()]
* `max_d`: See documentation for [ALE()]
* `requested_x_cols`: See documentation for [ALE()]. Note, however, that `ALEPlots` does not store `ordered_x_cols`.
}\if{html}{\out{</div>}}

}
}
}

\section{Consolidation of factors}{


When a categorical (unordered factor) variable has too many levels, the 1D column charts used to plot their ALE become unwieldy and hard to read. So, the \code{consolid_cats} argument sets a maximum number of categories (factor levels) to display. For example, for the default \code{consolid_cats = 10}:
\itemize{
\item If there are \code{consolid_cats} (default 10) or fewer categories, then there is no consolidation.
\item With more than \code{consolid_cats} categories, all categories are then ranked by decreasing absolute ALE \code{y} value. The top \code{consolid_cats - 1} (default 9) categories are retained.
\item All other categories are consolidated into one "other" category that reports the count of consolidated categories. Consolidated means are the weighted mean of all categories; consolidated medians, maximums, and minimums are the medians, maximums, and minimums, respectively, of all categories.
}

Sometimes, we have specific factor levels that we always want to see; we don't want them consolidated even if their effects are very low. In that case, see the argument specification for how to list such levels that must always be included.

Note that this consolidation is purely for visualization; the underlying ALE data is not consolidated. Only unordered factors are consolidated thus; ordered factors are never consolidated since their order is meaningful. Moreover, for now, only 1D ALE plots are consolidated.
}

\examples{
# See examples with ALE() and ModelBoot() objects.

}
