% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browse_ale}
\alias{browse_ale}
\title{Interactively browse an \code{ale} object with a Shiny app}
\usage{
browse_ale(id, obj = NULL, obj_name = NULL, ale_dataset = NULL)
}
\arguments{
\item{id}{character length 1. A required unique identifier to distinguish the
current module from any other modules in the same page or namespace. If two
instances of \code{browse_ale()} are called in the same page or namespace, they might
not work properly.}

\item{obj}{\code{ale_data} object. The \code{ale_data} object to browse. If NULL (default),
the app lets the user load a serialized object from the "Load" tab.}

\item{obj_name}{character length 1. An optional label with which \code{obj} will
be named in the app.}

\item{ale_dataset}{dataframe. The dataset on which \code{obj} was produced. If the
full dataset is unavailable, at least a random sample of the dataset should be
provided. When provided, this optional argument is used for some visualizations,
particularly for rug plots.}
}
\value{
This function returns a Shiny app that browses ale objects. It is designed such
that it can be either run standalone or embedded within Quarto or R Markdown
documents.
}
\description{
Interactively browse an \code{ale} object with a Shiny app.
}
\examples{
\donttest{
browse_ale()
}


}
