% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ale_ixn.R
\name{calc_ale_ixn}
\alias{calc_ale_ixn}
\title{Calculate ALE interaction data}
\usage{
calc_ale_ixn(X, model, x1_col, x2_col, y_cats, pred_fun, pred_type, max_x_int)
}
\arguments{
\item{X}{dataframe. Data for which ALE is to be calculated. The y (outcome) column is absent.}

\item{model}{See documentation for \code{\link[=ale]{ale()}}}

\item{x1_col, x2_col}{character(1),character(1). Name of single columns in X for which ALE interaction data is to be calculated. \code{x1_col} can be of any standard datatype (logical, factor, or numeric) but \code{x2_col} can only be numeric.}

\item{pred_fun}{See documentation for \code{\link[=ale]{ale()}}}

\item{pred_type}{See documentation for \code{\link[=ale]{ale()}}}

\item{max_x_int}{See documentation for \code{\link[=ale]{ale()}}}
}
\description{
This function is not exported. It is adapted from \code{\link[ALEPlot:ALEPlot]{ALEPlot::ALEPlot()}}.
This function is not usually called directly by the user. For details about arguments not documented here, see \code{\link[=ale]{ale()}}.
}
\details{
@author Dan Apley (source of original calculation of ALE in \code{\link[ALEPlot:ALEPlot]{ALEPlot::ALEPlot()}})
@references Apley, Daniel W., and Jingyu Zhu. "Visualizing the effects of predictor variables in black box supervised learning models." Journal of the Royal Statistical Society Series B: Statistical Methodology 82.4 (2020): 1059-1086.
@author Chitu Okoli (completely rewrote the code based on \code{ALEPlot::ALEPlot()})
}
