% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calc_p}
\alias{calc_p}
\title{S7 generic for calculating p-values for an object}
\usage{
calc_p(obj, ...)
}
\arguments{
\item{obj}{object.}

\item{...}{Arguments to pass to the implemented method.}
}
\description{
Calculate p-values
}
