% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_stats}
\alias{calc_stats}
\title{Calculate statistics from ALE y values.}
\usage{
calc_stats(y, bin_n, y_vals = NULL, ale_y_norm_fun = NULL, x_type = "numeric")
}
\arguments{
\item{y}{numeric. Vector of ALE y values.}

\item{bin_n}{numeric. Vector of counts of rows in each ALE bin. Must be the same length as \code{y}.}

\item{y_vals}{numeric. Entire vector of y values. Needed for normalization. If not provided, ale_y_norm_fun must be provided.}

\item{ale_y_norm_fun}{function. Result of \code{create_ale_y_norm_function()}. If not provided, \code{y_vals} must be provided. \code{calc_stats()} could be faster if \code{ale_y_norm_fun} is provided, especially in bootstrap workflows that call the same function many, many times.}

\item{x_type}{character(1). Datatype of the x variable on which the ALE y is based. Values are the result of \code{var_type()}. Used to determine how to correctly calculate ALE, so if the value is not the default \code{"numeric"}, then it must be set correctly.}
}
\value{
Named numeric vector:
\itemize{
\item aled: ALE deviation (ALED)
\item aler_min: Minimum (lower value) of the ALE range (ALER)
\item aler_max: Maximum (upper value) of the ALE range (ALER)
\item naled: Normalized ALE deviation (ALED)
\item naler_min: Normalized minimum (lower value) of the ALE range (ALER)
\item naler_max: Normalized maximum (upper value) of the ALE range (ALER)
}
}
\description{
Not exported. The following statistics are calculated based on a vector of ALE y values:
}
\details{
\itemize{
\item ALE deviation (ALED)
\item ALE range (ALER): range from minimum value of any ALE \code{y} to the maximum value of any \code{y}. This is a very simple indication of the dispersion in ALE \code{y} values.
\item Normalized ALE deviation (NALED)
\item Normalized ALE range (NALER)
}

Note that if any ALE \code{y} values are missing, they will be deleted from the calculation (with their corresponding bin_n).
}
