% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_sums}
\alias{col_sums}
\title{Sum up a matrix across columns}
\usage{
col_sums(mx, na.rm = FALSE, dims = 1)
}
\arguments{
\item{mx}{numeric matrix}

\item{na.rm}{logical(1). TRUE if missing values (\code{NA}) should be ignored in the summation. If FALSE (default), even one missing value will result in \code{NA} for the entire column.}

\item{dims}{See documentation for \code{\link[base:colSums]{base::colSums()}}}
}
\value{
numeric vector whose length is number of columns of \code{mx}, whose values are the sums of each column of \code{mx}.
}
\description{
Adaptation of \code{\link[base:colSums]{base::colSums()}} that, when all values in a column are NA, sets the sum to NA rather than zero as \code{\link[base:colSums]{base::colSums()}} does. Calls \code{\link[base:colSums]{base::colSums()}} internally.
}
\examples{
set.seed(1)
mx <- matrix(
  sample(1:6, 24, replace = TRUE),
  nrow = 4
)
# Randomly set some values as missing
mx[sample(1:24, 12)] <- NA
mx
col_sums(mx, na.rm = FALSE)
col_sums(mx, na.rm = TRUE)

}
