% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{extract_2D_diags}
\alias{extract_2D_diags}
\title{Extract all NWSE diagonals from a matrix}
\usage{
extract_2D_diags(mx)
}
\arguments{
\item{mx}{matrix}
}
\value{
A list whose elements each represent one diagonal of \code{mx}. Each diagonal element is a list of two elements: \code{coords} is a numeric vector pair of row-column coordinates; \code{values} is the value of the diagonal at the coordinate give by \code{coords}.
}
\description{
Extracts all diagonals from a matrix in the NWSE direction (upper left down to lower right).
}
