% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{extract_3D_diags}
\alias{extract_3D_diags}
\title{Extract all FNWBSE diagonals from a 3D array}
\usage{
extract_3D_diags(ray)
}
\arguments{
\item{ray}{a 3-dimensional array}
}
\value{
A list whose elements each represent one diagonal of \code{ray}. Each diagonal element is a list of two elements: \code{origin} is the 3D coordinates (row, column, depth) of the first element of the diagonal; \code{values} is a vector of the diagonal that starts from \code{origin}.
}
\description{
Extracts all diagonals from a 3D array in the FNWBSE direction (front upper left down to back lower right).
}
