% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{get-ALEPlots-method}
\alias{get-ALEPlots-method}
\title{get method for ALEPlots objects}
\value{
A list of \code{ggplot} objects using the category and dimension
conventions described for the \link[=get-ALE-method]{ALE method}. This differs
from the \link[=subset-ALEPlots-method]{subset() method}, which returns an
\code{ALEPlots} object.
}
\description{
Retrieve specific plots from an \code{ALEPlots} object with the \code{\link[=get]{get()}} generic.
Unlike the \link[=subset-ALEPlots-method]{subset() method}, which returns an
\code{ALEPlots} object, this method returns a list of \code{ggplot2::ggplot} objects.

See the \link[=get-ALE-method]{get() method for ALE objects} for arguments and
output structure not described here.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{get(obj, x_cols = NULL, ..., exclude_cols = NULL, type = "ale", cats = NULL, simplify = TRUE, silent = FALSE)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{obj}: ALEPlots object from which to retrieve ALE elements.
\item \code{type}: character(1). What type of ALEPlots to retrieve: \code{'ale'} for standard ALE plots or \code{'effect'} for ALE effects plots. See \code{cats} argument for options for categorical plots.
\item \code{cats}: character. The categories (one or more) of a categorical outcome variable to retrieve. To retrieve all categories as individual category plots, leave \code{cats} at the default \code{NULL}. For categorical plots that combine all categories, specify \code{cats = ".all"}. (Don't forget the "." in ".all", which avoids naming conflicts with legitimate categories that might be named "all".) For such all-category plots, \code{type} must be set to "overlay" or "facet" for the specific desired type of categorical plot.
}
}

