% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBoot_methods.R
\name{get-ModelBoot-method}
\alias{get-ModelBoot-method}
\title{get method for ModelBoot objects}
\value{
See the \link[=get-ALE-method]{get() method for ALE objects}.
}
\description{
Retrieve specific ALE elements from a \code{ModelBoot} object with the \code{\link[=get]{get()}}
generic. This method is similar to the \link[=get-ALE-method]{ALE method}, except
that the user may specify what \code{type} of ALE data to retrieve.

See the \link[=get-ALE-method]{get() method for ALE objects} for arguments and
output structure not described here.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{get(obj, x_cols = NULL, what = "ale", ..., exclude_cols = NULL, type = "auto", stats = NULL, cats = NULL, ale_centre = "median", simplify = TRUE)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{obj}: ModelBoot object from which to retrieve ALE elements.
\item \code{type}: character(1). The type of ModelBoot ALE elements to retrieve: \code{'single'} for the ALE calculated on the full data set or \code{'boot'} for the bootstrapped ALE data (based on full-model bootstrapping). The default \code{'auto'} will retrieve \code{'boot'} if it is available and \code{'single'} otherwise.
}
}

