% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get.calc_p}
\alias{get.calc_p}
\title{Calculate and append p-values to an ALE object}
\arguments{
\item{obj}{object.}

\item{...}{Arguments to pass to the implemented method.}
}
\value{
\code{obj} with p-values included in its statistics.
}
\description{
Calculating p-values for ALE objects is usually very slow, so it is not done by default. the \code{calc_p()} method receives an \code{ALEpDist} object, uses it to calculate p-values, and then appends these p-values to the \code{ALE} object.
}
