% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ale.R
\name{idxs_kolmogorov_smirnov}
\alias{idxs_kolmogorov_smirnov}
\title{Sorted categorical indices based on Kolmogorov-Smirnov distances for empirically ordering categorical categories.}
\usage{
idxs_kolmogorov_smirnov(X, x_col, n_bins, x_int_counts)
}
\description{
Sorted categorical indices based on Kolmogorov-Smirnov distances for empirically ordering categorical categories.
}
