% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{intrapolate_3D}
\alias{intrapolate_3D}
\title{Intrapolate missing values of a 3D array}
\usage{
intrapolate_3D(ray, consolidate = TRUE)
}
\arguments{
\item{ray}{numeric array of three dimensions.}

\item{consolidate}{logical(1). See return value.}
}
\value{
If \code{consolidate = TRUE} (default), returns a numeric array of the same dimensions as the input \code{ray} with internal missing values linearly intrapolated. If \code{consolidate = FALSE}, returns a list of intrapolations for missing values from each slice and diagonal direction.
}
\description{
This intrapolation algorithm replaces internal missing values in a three-dimensional array. For how it works, see the details of \code{\link[=intrapolate_2D]{intrapolate_2D()}}. Based on that,  \code{\link[=intrapolate_3D]{intrapolate_3D()}} does the following:
\itemize{
\item Slice the 3D array into 2D matrices along the rows, columns, and depth dimensions.
\item Use \code{\link[=intrapolate_2D]{intrapolate_2D()}} to calculate 2D intrapolations based on the algorithm of \code{\link[=intrapolate_1D]{intrapolate_1D()}}. See details in their documentation.
\item In addition, calculate intrapolations along the four directions of 3D diagonals: front northwest to back southeast, that is, front upper left down to back lower right  (FNWBSE), FSWBNE, FSEBNW, and FNEBSW.
\item The 3D intrapolation is the mean intrapolation from any of these 2D or 3D values. In taking the mean, missing intrapolations are removed.
\item When there is no intrapolation available from any of the directions, the missing value remains missing.
}
}
