% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_stats.R
\name{mad}
\alias{mad}
\title{Mean absolute deviation}
\usage{
mad(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector. Values for which to calculate the mean absolute deviation.}

\item{na.rm}{single logical. TRUE if missing values should be removed; FALSE
if they should be retained.}
}
\value{
MAD of the \code{x} values. If any value of \code{x} is \code{NA} and \code{na.rm = FALSE},
then returns \code{NA}.
}
\description{
Returns the mean absolute deviation (MAD) of values relative to their mean.
}
\details{
Note: This function is different from \code{mad} in the \code{stats} package, which returns the median
absolute deviation relative to the median, a different statistic. If you want
to access that function, you can explicitly refer to it as \code{stats::mad}.
}
