% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{params_data}
\alias{params_data}
\title{Improvements:
\itemize{
\item Validation: ensure that the object is atomic (not just a vector)
\item For factors, get all classes and check if any class_x is a factor or ordered
\item Add arguments to return a unique mode with options to sort: occurrence order, lexigraphical
}}
\usage{
params_data(data, data_name = var_name(data), sample_size = 500, seed = 0)
}
\description{
Improvements:
\itemize{
\item Validation: ensure that the object is atomic (not just a vector)
\item For factors, get all classes and check if any class_x is a factor or ordered
\item Add arguments to return a unique mode with options to sort: occurrence order, lexigraphical
}
}
