% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{plot-ALE-method}
\alias{plot-ALE-method}
\title{plot method for \code{ALE} objects}
\description{
This \code{\link[graphics:plot]{graphics::plot()}} method simply calls the constructor for an
\code{ALEPlots} object.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{plot(x, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: ALE object.
\item \code{...}: Arguments passed to \code{\link[=ALEPlots]{ALEPlots()}}
}
}

