% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{plot.ALE}
\alias{plot.ALE}
\title{plot method for \code{ALE} objects}
\usage{
## S7 method for class <ale::ALE>
plot(x, ...)
}
\arguments{
\item{x}{ALE object.}

\item{...}{Arguments passed to \code{\link[=ALEPlots]{ALEPlots()}}}
}
\description{
This plot method simply calls the constructor for an \code{ALEPlots} object.
}
