% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ale_boot}
\alias{plot.ale_boot}
\title{plot method for \code{ale_boot} objects}
\usage{
\method{plot}{ale_boot}(ale_boot_obj, type = "ale")
}
\arguments{
\item{ale_boot_obj}{ale_boot object. The object of class \code{ale_boot} containing data to be plotted.}

\item{character(1).}{'ale' for regular ALE plots; 'effects' for an ALE statistic effects plot.}
}
\description{
plot method for \code{ale_boot} objects
}
