% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ale_obj}
\alias{plot.ale_obj}
\title{plot method for ALE objects}
\usage{
\method{plot}{ale_obj}(ale_obj, y_type, y_summary, ..., x_y_data = NULL, compact_plots = FALSE)
}
\arguments{
\item{ale_obj}{ale object. The object of class \code{ale_obj} containing data to be plotted.}

\item{y_type}{character(1). The type of the response variable.}

\item{y_summary}{matrix. Summary statistics for the response variable.}

\item{...}{not used. Enforces explicit naming of subsequent arguments.}

\item{x_y_data}{Data for the x and y variables.}

\item{compact_plots}{Logical indicating if the plots should be compact.}
}
\description{
plot method for ALE objects
}
