% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{plot.ALEPlots}
\alias{plot.ALEPlots}
\title{Plot method for ALEPlots object}
\usage{
## S7 method for class <ale::ALEPlots>
plot(x, max_print = 20L, ...)
}
\arguments{
\item{x}{An object of class \code{ALEPlots}.}

\item{max_print}{integer(1). The maximum number of plots that may be printed at a time. 1D plots and 2D are printed on separate pages, so this maximum applies separately to each dimension of ALE plots, not to all dimensions combined.}

\item{...}{Arguments to pass to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plot an \code{ALEPlots} object.
}
