% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBoot_methods.R
\name{plot.ModelBoot}
\alias{plot.ModelBoot}
\title{plot method for \code{ModelBoot} objects}
\usage{
## S7 method for class <ale::ModelBoot>
plot(x, ...)
}
\arguments{
\item{x}{ModelBoot object.}

\item{...}{Arguments passed to \code{\link[=ALEPlots]{ALEPlots()}}}
}
\description{
This plot method simply calls the constructor for an \code{ALEPlots} object.
}
