% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_core.R
\name{plot_ale}
\alias{plot_ale}
\title{Plot ALE data}
\usage{
plot_ale(
  ale_data,
  x_col,
  y_col,
  y_type,
  y_summary,
  relative_y = "median",
  plot_alpha = 0.05,
  data = NULL,
  rug_sample_size = 500,
  min_rug_per_interval = 1,
  seed = 0
)
}
\arguments{
\item{ale_data}{tibble. Output data from \code{calc_ale}.}

\item{x_col}{character length 1. Name of single column whose ALE data is to
be plotted.}

\item{y_col}{character length 1. Name of y (output) column whose ALE data is to
be plotted.}

\item{y_type}{See documentation for \code{ale}}

\item{y_summary}{named double. Named vector of y summary statistics to be used
for plotting.}

\item{relative_y}{See documentation for \code{ale}}

\item{plot_alpha}{See documentation for \code{ale}}

\item{data}{dataframe. If provided, used to generate rug plots. Must at least
contain columns x_col and y_col; any other columns are not used.}

\item{rug_sample_size, min_rug_per_interval}{See documentation for \code{ale}}

\item{seed}{See documentation for \code{ale}}
}
\description{
This function is not exported. It creates a ggplot object that plots the input
ALE data generated from \code{calc_ale}.
This function is not usually called directly by the user. For details about
arguments not documented here, see \code{ale}.
}
