% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_var_for_ale.R
\name{prep_var_for_ale}
\alias{prep_var_for_ale}
\title{Compute preparatory data for ALE calculation}
\usage{
prep_var_for_ale(x_col, x_type, x_vals, ale_x, ale_n, max_x_int, X = NULL)
}
\arguments{
\item{x_col}{character(1). Name of single column in X for which ALE data is to be calculated.}

\item{x_type}{character(1). var_type() of x_col.}

\item{x_vals}{vector. The values of x_col.}

\item{ale_x, ale_n}{See documentation for \code{\link[=calc_ale]{calc_ale()}}}

\item{max_x_int}{See documentation for \code{\link[=ale]{ale()}}}

\item{X}{See documentation for \code{\link[=calc_ale]{calc_ale()}}. Used only for categorical x_col.}
}
\description{
This function is not exported. It computes data needed to calculate its ALE values.
}
