% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{print-ALE-method}
\alias{print-ALE-method}
\title{print Method for ALE object}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ALE object with the \code{\link[base:print]{base::print()}} generic.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{print(x, details = TRUE, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: An object of class \code{ALE}.
\item \code{details}: logical(1). If \code{TRUE} (default), all brief details are printed. If \code{FALSE}, only minimal information is printed.
\item \code{...}: Additional arguments (currently not used).
}
}

\examples{
\donttest{
lm_cars <- stats::lm(mpg ~ ., mtcars)
ale_cars <- ALE(lm_cars, p_values = NULL)
print(ale_cars)
}

}
