% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{print-ALEPlots-method}
\alias{print-ALEPlots-method}
\title{Print method for ALEPlots object}
\value{
Invisibly returns \code{x}.
}
\description{
Print an \code{ALEPlots} object with the \code{\link[base:print]{base::print()}} generic by calling its
\code{\link[graphics:plot]{graphics::plot()}} method.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{print(x, max_print = 20L, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: An object of class \code{ALEPlots}.
\item \code{max_print}: See the \link[=plot-ALEPlots-method]{plot() method}.
\item \code{...}: Additional arguments (currently not used).
}
}

