% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBoot_methods.R
\name{print-ModelBoot-method}
\alias{print-ModelBoot-method}
\title{print method for ModelBoot object}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ModelBoot object with the \code{\link[base:print]{base::print()}} generic.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{print(x, details = TRUE, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: An object of class \code{ModelBoot}.
\item \code{details}: logical(1). If \code{TRUE} (default), all brief details are printed. If \code{FALSE}, only minimal information is printed.
\item \code{...}: Additional arguments (currently not used).
}
}

\examples{
\donttest{
lm_cars <- stats::lm(mpg ~ wt + gear, mtcars)
mb <- ModelBoot(lm_cars, boot_it = 2, ale_p = NULL)
print(mb)
}

}
