% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print.ale_plots}
\alias{print.ale_plots}
\title{Print method for ale_plots object}
\usage{
\method{print}{ale_plots}(x, max = 20L, ...)
}
\arguments{
\item{x}{An object of class \code{ale_plots}.}

\item{max}{See documentation for \code{\link[=plot.ale_plots]{plot.ale_plots()}}}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ale_plots object by calling plot().
}
