% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{print.ALEPlots}
\alias{print.ALEPlots}
\title{Print method for ALEPlots object}
\usage{
## S7 method for class <ale::ALEPlots>
print(x, max_print = 20L, ...)
}
\arguments{
\item{x}{An object of class \code{ALEPlots}.}

\item{max_print}{See documentation for \code{\link[=plot.ALEPlots]{plot.ALEPlots()}}}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ALEPlots object by calling plot().
}
