% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBoot_methods.R
\name{print.ModelBoot}
\alias{print.ModelBoot}
\title{print method for ModelBoot object}
\usage{
## S7 method for class <ale::ModelBoot>
print(x, details = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{ModelBoot}.}

\item{details}{logical(1). If \code{TRUE} (default), all brief details are printed. If \code{FALSE}, only minimal information is printed.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a ModelBoot object.
}
\examples{
\donttest{
lm_cars <- stats::lm(mpg ~ wt + gear, mtcars)
mb <- ModelBoot(lm_cars, boot_it = 2, ale_p = NULL)
print(mb)
}

}
