% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy.R
\name{mae}
\alias{mae}
\alias{rmse}
\alias{mad}
\alias{auc}
\title{Mean absolute error}
\usage{
mae(actual, pred, na.rm = FALSE)

rmse(actual, pred, na.rm = FALSE)

mad(x, na.rm = FALSE, version = "mean", ...)

auc(actual, pred, na.rm = FALSE, binary_true_value = NULL)
}
\arguments{
\item{actual}{any atomic vector. Actual label values from a dataset. They must be binary; that is, there must be exactly two distinct values (other than missing values, which are allowed). If the values are not naturally converted to logical \code{TRUE} and \code{FALSE}, then specify which of the two values should be considered \code{TRUE} with the argument \code{binary_true_value}.}

\item{pred}{numeric vector. Predictions corresponding to each respective
element in \code{actual}.}

\item{na.rm}{single logical. TRUE if missing values should be removed; FALSE
if they should be retained. If TRUE, then if any element of either actual or
pred is missing, its paired element will be also removed.}

\item{x}{numeric vector. Values for which to calculate the mean absolute deviation.}

\item{version}{character(1). By default (\code{version = 'mean'}), \code{mad()} returns the mean absolute deviation (MAD) of values relative to their mean. If \code{version = 'median'}, it returns the \code{stats::mad()} function, the median absolute deviation relative to their median. Any other value gives an error.}

\item{...}{Arguments to pass to \code{stats::mad()} if \code{version = 'median'}. See the \code{version} argument for details.}

\item{binary_true_value}{any single atomic value. The value of \code{actual} that is considered \code{TRUE}; any other value of \code{actual} is considered \code{FALSE}. For example, if 2 means \code{TRUE} and 1 means \code{FALSE}, then set \code{binary_true_value} as \code{2}.}
}
\value{
MAE of \code{actual} and \code{pred}. If any value in \code{actual} or \code{pred} is
\code{NA} and \code{na.rm = FALSE}, then returns \code{NA}.

RMSE of \code{actual} and \code{pred}. If any value in \code{actual} or \code{pred} is \code{NA} and \code{na.rm = FALSE}, returns \code{NA}.

MAD of the \code{x} values. If any value of \code{x} is \code{NA} and \code{na.rm = FALSE}, returns \code{NA}.

MAE of \code{actual} and \code{pred}. If any value in \code{actual} or \code{pred} is
\code{NA} and \code{na.rm = FALSE}, then returns \code{NA}.
}
\description{
Returns the mean absolute error (MAE) of predicted values relative to the actual values.

Returns the root mean squared error (RMSE) of predicted values relative to the actual values.

Returns the mean absolute deviation (MAD) of values relative to their mean.

Returns the area under the ROC curve based on comparing the predicted probabilities to the actual binary values.
}
