% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_stats.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(actual, predicted, na.rm = FALSE)
}
\arguments{
\item{actual}{numeric vector. Actual (true) values from a dataset.}

\item{predicted}{numeric vector. Predictions corresponding to each respective
element in \code{actual}.}

\item{na.rm}{single logical. TRUE if missing values should be removed; FALSE
if they should be retained. If TRUE, then if any element of either actual or
predicted is missing, its paired element will be also removed.}
}
\value{
RMSE of \code{actual} and \code{predicted}. If any value in \code{actual} or \code{predicted} is
\code{NA} and \code{na.rm = FALSE}, then returns \code{NA}.
}
\description{
Returns the root mean squared error (RMSE) of predicted values relative to the
actual values.
}
