% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy.R
\name{standardized_accuracy}
\alias{standardized_accuracy}
\title{Create standardized accuracy (staccuracy) functions.}
\usage{
standardized_accuracy(error_fun, ref_fun)
}
\arguments{
\item{error_fun}{function. The unquoted function name of the function that calculates error (or accuracy) measure. This function must be of the signature \code{error_fun(actual, pred, na.rm = FALSE)} where \code{actual} and \code{pred} are required numeric vectors of the true (actual) labels and the predicted estimates, respectively; \code{na.rm} is an optional scalar logical that indicates if NA values should be removed (\code{TRUE}) or not (\code{FALSE}, default).}

\item{ref_fun}{function. The unquoted function name of the function that calculates the reference error, accuracy, or deviation measure. This function must be of the signature \code{ref_fun(actual, na.rm = FALSE)} where \code{actual} and \code{na.rm} have the same meanings as described for \code{error_fun}.}
}
\value{
Function with signature \verb{function(actual, pred, na.rm = FALSE)} whose arguments have the same meanings as  described for the \code{error_fun} input argument.
}
\description{
Create standardized accuracy (staccuracy) functions.
}
