% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{subset-ALEPlots-method}
\alias{subset-ALEPlots-method}
\title{subset method for ALEPlots object}
\value{
An \code{ALEPlots} object reduced to the variables and interactions
specified by \code{x_cols} and \code{exclude_cols}. This differs from the
\link[=get-ALEPlots-method]{get() method}, which returns a list of \code{ggplot}
objects and loses the special \code{ALEPlots} behaviour like plotting, printing,
and summarizing multiple plots.
}
\description{
Use the \code{\link[base:subset]{base::subset()}} generic on an \code{ALEPlots} object to produce another
\code{ALEPlots} object containing only the requested variables and interactions.

See the \link[=get-ALE-method]{get() method for ALE objects} for shared argument
conventions.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{subset(x, x_cols = NULL, ..., exclude_cols = NULL, include_eff = TRUE, silent = FALSE)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: An object of class \code{ALEPlots}.
\item \code{x_cols}, \code{exclude_cols}: See the \link[=get-ALE-method]{get() method for ALE objects}.
\item \code{...}: not used. Inserted to require explicit naming of subsequent arguments.
\item \code{include_eff}: logical(1). \code{x_cols} and \code{exclude_cols} specify precisely which variables to include or exclude in the subset. However, multivariable plots like ALE effects plot are ambiguous because they cannot be subsetted to remove some existing variables. \code{include_eff = TRUE} (default) includes the ALE effects plot in the subset rather than dropping it, if it is available.
\item \code{silent}: See documentation for \code{\link[=ALE]{ALE()}}
}
}

