% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{summary-ALEPlots-method}
\alias{summary-ALEPlots-method}
\title{summary method for ALEPlots object}
\value{
Summary string.
}
\description{
Use the \code{\link[base:summary]{base::summary()}} generic to return concise information about an
\code{ALEPlots} object.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{summary(object, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{object}: An object of class \code{ALEPlots}.
\item \code{...}: Not used
}
}

