% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_core.R
\name{var_type}
\alias{var_type}
\title{Determine the datatype of a vector}
\usage{
var_type(var)
}
\arguments{
\item{var}{vector whose datatype is to be determined}
}
\value{
Returns generic datatypes of R basic vectors according to the following mapping:
\itemize{
\item \code{logical} returns 'binary'
\item \code{numeric} values (e.g., \code{integer} and \code{double}) return 'numeric'
\item However, if the only values of numeric are 0 and 1, then it returns 'binary'
\item unordered \code{factor} returns 'multinomial'
\item \code{ordered} \code{factor} returns 'ordinal'
}
}
\description{
Not exported. See @returns for details of what it does.
}
