% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy.R
\name{winsorize}
\alias{winsorize}
\alias{win_mae}
\alias{win_rmse}
\title{Winsorize a numeric vector.}
\usage{
winsorize(x, win_range)

win_mae(actual, pred, win_range = range(actual), na.rm = FALSE)

win_rmse(actual, pred, win_range = range(actual), na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector.}

\item{win_range}{numeric(2). The minimum and maximum allowable values for the \code{pred} predictions. Defaults to the minimum and maximum values of the provided \code{actual} values.}

\item{actual}{numeric vector. Actual (true) values from a dataset.}

\item{pred}{numeric vector. Predictions corresponding to each respective
element in \code{actual}.}

\item{na.rm}{single logical. \code{TRUE} if missing values should be removed; \code{FALSE} if they should be retained. If \code{TRUE}, then if any element of either \code{actual}  or \code{pred} is missing, its paired element will be also removed.}
}
\value{
Winsorized MAE of \code{actual} and \code{pred}. See mae() for details.

Winsorized MAE of \code{actual} and \code{pred}. See mae() for details.
}
\description{
Returns the input vector values with values less than or greater than the provided minimum or maximum replaced by the provided minimum or maximum, respectively.

Returns the mean absolute error (MAE) with predictions winsorized within a specified range.
}
