% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ale.R
\name{calc_ale}
\alias{calc_ale}
\title{Calculate ALE data}
\usage{
calc_ale(
  data,
  model,
  x_cols,
  y_col,
  y_cats,
  pred_fun,
  pred_type,
  max_num_bins,
  boot_it,
  seed,
  boot_alpha,
  boot_centre,
  boot_ale_y = FALSE,
  .bins = NULL,
  ale_y_norm_funs = NULL,
  p_dist = NULL
)
}
\arguments{
\item{data}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{model}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{x_cols}{character(1 or 2). Names of columns in X for which ALE data is to be calculated. Length 1 for 1D ALE and length 2 for 2D ALE.}

\item{y_col}{character(1). Name of the target y column.}

\item{y_cats}{character. The categories of y. For most cases with non-categorical y, \code{y_cats == y_col}.}

\item{pred_fun}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{pred_type}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{max_num_bins}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{boot_it}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{seed}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{boot_alpha}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{boot_centre}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{boot_ale_y}{logical(1). If \code{TRUE}, return the bootstrap matrix of ALE y values. If \code{FALSE} (default) return NULL for the \code{boot_ale_y} element of the return value.}

\item{.bins}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{ale_y_norm_funs}{list of functions. Custom functions for normalizing ALE y for statistics. It is usually a list(1), but for categorical y, there is a distinct function for each y category. If provided, ale_y_norm_funs saves some time since it is usually the same for all all variables throughout one call to \code{\link[=ALE]{ALE()}}. For now, used as a flag to determine whether statistics will be calculated or not; if NULL, statistics will not be calculated.}

\item{p_dist}{See documentation for \code{p_values} in \code{\link[=ALE]{ALE()}}}
}
\description{
This function is not exported. It is a complete reimplementation of the ALE algorithm relative to the reference in \code{\link[ALEPlot:ALEPlot]{ALEPlot::ALEPlot()}}. In addition to adding bootstrapping and handling of categorical y variables, it reimplements categorical x interactions.
}
\details{
For details about arguments not documented here, see \code{\link[=ALE]{ALE()}}.
}
\references{
Apley, Daniel W., and Jingyu Zhu. "Visualizing the effects of predictor variables in black box supervised learning models." Journal of the Royal Statistical Society Series B: Statistical Methodology 82.4 (2020): 1059-1086.

Okoli, Chitu. 2023. “Statistical Inference Using Machine Learning and Classical Techniques Based on Accumulated Local Effects (ALE).” arXiv. doi:10.48550/arXiv.2310.09877.
}
