% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get.ALE}
\alias{get.ALE}
\title{get method for ALE objects}
\arguments{
\item{obj}{ALE object from which to retrieve elements.}

\item{x_cols, exclude_cols}{character, list, or formula. Columns names and interaction terms from \code{obj} requested in one of the special \code{x_cols} formats. The default value of \code{NULL} for \code{x_cols} retrieves all available data of the output requested in \code{what}. See details in the documentation for \code{\link[=resolve_x_cols]{resolve_x_cols()}}.}

\item{what}{character(1). What kind of output is requested. Must be one (and only one) of \code{c('ale', 'boot_data')}. Default is \code{'ale'}. If \code{stats} is specified and \code{what = 'ale'}, then ALE statistics are retrieved. Otherwise, \code{get()} errors if \code{stats} is specified and \code{what} has some other value.}

\item{...}{not used. Inserted to require explicit naming of subsequent arguments.}

\item{stats}{character(1). Retrieve statistics. If \code{stats} is specified, then \code{what} must be \code{'ale'}. See the return value details below for valid values for \code{stats}.}

\item{cats}{character. Optional category names to retrieve if the ALE is for a categorical y outcome model.}

\item{simplify}{logical(1). If \code{TRUE} (default), the results will be simplified to the simplest structure possible to give the requested results.}

\item{silent}{See documentation for \code{\link[=resolve_x_cols]{resolve_x_cols()}}}
}
\value{
Regardless of the requested data, all \code{\link[=get.ALE]{get.ALE()}} have a common structure:
\itemize{
\item If more than one category of the y outcome is returned, then the top level is a list named by each category. However, the y outcome is not categorical or only one category of multiple possibilities is specified using the \code{cats} argument, then the top level never has categories, regardless of the value of \code{simplify}.
\item The next level (or top level if there are zero or one category) is a list with one or two levels:
\itemize{
\item \code{d1}: 1D ALE elements.
\item \code{d2}: 2D ALE elements.
However, if elements of only one dimension (either 1D or 2D) are requested and \code{simplify = TRUE} (default), the empty list is eliminated and the level is skipped to provide only the elements present. For example, if only 1D ALE data is requested, then there will be no \code{d1} sublist but only a list of the ALE data as described for the next level. If \code{simplify = FALSE}, both \code{d1} and \code{d2} sublists will always be returned; the empty sublist will be \code{NULL}.
}
\item For the \code{d1} sublist, there is no further hierarchy: the returned data is as described below. For the \code{d2} sublist representing \code{var1} by \code{var2} interactions, the next level is a named list of all \code{var1} elements. This level in turn consists of a named list of \code{var2} elements with the actual data as described below. For example, for ALE data, \code{result <- get(obj, list(d1 = "var1", d2 = list(c("var1", "var2"))))} would return \code{result} as a list where the 1D ALE data is in \code{result$d1$var1} and the 2D ALE data in \code{result$d2$var1$var2}.
}

While all results follow the general structure just described, the specific type of data returned depends on the values of the \code{what} and \code{stats} arguments:
\describe{
\item{\code{what = 'ale'} (default) and \code{stats = NULL} (default)}{A list whose elements, named by each requested x variable, are each a tibble. The rows each represent one ALE bin. The tibble has the following columns:
* \code{var.bin} or \code{var.ceil} where \code{var} is the name of a variable (column): For non-numeric x, \code{var.bin} is the value of each of the ALE categories. For numeric x, \code{var.ceil} is the value of the upper bound (ceiling) of each ALE bin. The first "bin" of numeric variables represents the minimum value. For 2D ALE with an \code{var1} by \code{var2} interaction, both \code{var1.bin} and \code{var2.bin} columns are returned (or \code{var1.ceil} or \code{var2.ceil}for numeric \code{var1} or \code{var2}).
* \code{.n}: the number of rows of data in each bin represented by \code{var.bin} or \code{var.ceil}. For numeric x, the first bin contains all data elements that have exactly the minimum value of x. This is often 1, but might be more than 1 if more than one data element has exactly the minimum value.
* \code{.y}: the ALE function value calculated for that bin. For bootstrapped ALE, this is the same as \code{.y_mean} by default or \code{.y_median} if \code{boot_centre = 'median'}. Regardless, both \code{.y_mean} and \code{.y_median} are returned as columns here.
* \code{.y_lo}, \code{.y_hi}: the lower and upper confidence intervals, respectively, for the bootstrapped \code{.y} value based on the \code{boot_alpha} argument in the \code{\link[=ALE]{ALE()}} constructor.
}

\item{\code{what = 'boot_data'} and \code{stats = NULL} (default)}{A list whose elements, named by each requested x variable, are each a tibble. These are the data from which \code{.y_mean}, \code{.y_median}, \code{.y_lo}, and \code{.y_hi} are summarized when \code{what = 'ale'}. The rows each represent one ALE bin for a specified bootstrap iteration. The tibble has the following columns:
* \code{.it}: The bootstrap iteration. Iteration 0 represents the ALE calculations on the full dataset; the remaining values of \code{.it} are from 1 to \code{boot_it} (number of bootstrap iterations specified in the \code{\link[=ALE]{ALE()}} constructor.
* \code{var} where \code{var} is the name of a variable (column): For non-numeric x, \code{var} is the value of each of the ALE categories. For numeric x, \code{var} is the value of the upper bound (ceiling) of each ALE bin. They are otherwise similar to their meanings described for \code{what = 'ale'} above.
* \code{.n} and \code{.y}: Same as for \code{what = 'ale'}.
}

\item{\code{what = 'ale'} (default) and \code{stats = 'estimate'}}{A list with elements \code{d1} and \code{d2} with the value of each ALE statistic. Each row represents one variable or interaction. The tibble has the following columns:
* \code{term} or \code{term1} and \code{term2}: The variable or column for the 1D (\code{term}) or 2D (\code{term1} by \code{term2}) ALE statistic.
* \code{aled}, \code{aler_min}, \code{aler_max}, \code{naled}, \code{naler_min}, \code{naler_max}: the respecitve ALE statistic for the variable or interaction.
}

\item{\code{what = 'ale'} (default) and \code{stats} is one value in \code{c('aled', 'aler_min', 'aler_max', 'naled', 'naler_min', 'naler_max')}}{A list with elements \code{d1} and \code{d2} with the distribution value of the single requested ALE statistic. Each element \code{d1} and \code{d2} is a tibble. Each row represents one variable or interaction. The tibble has the following columns:
* \code{term} or \code{term1} and \code{term2}: Same as for \code{stats = 'estimate'}.
* \code{estimate}, \code{mean}, \code{median}: The average of the bootstrapped value of the requested statistic. \code{estimate} is equal to either \code{mean} or \code{median} depending on the \code{boot_centre} argument in the \code{\link[=ALE]{ALE()}} constructor. If ALE is not bootstrapped, then \code{estimate}, \code{mean}, and \code{median} are equal.
* \code{conf.low}, \code{conf.high}: the lower and upper confidence intervals, respectively, for the bootstrapped statistic based on the \code{boot_alpha} argument in the \code{\link[=ALE]{ALE()}} constructor. If ALE is not bootstrapped, then \code{estimate}, \code{conf.low}, and \code{conf.high} are equal.
}

\item{\code{what = 'ale'} (default) and \code{stats = 'all'}}{A list with elements \code{d1} and \code{d2} with the distribution values of all available ALE statistics for the requested variables and interactions. Whereas the \code{stats = 'aled'} (for example) format returns data for a single statistic, \code{stats = 'all'} returns all statistics for the requested variables. Each element is a list with the requested \code{d1} and \code{d2} sub-elements as described in the general structure above. Each data element is a tibble. Each row represents one ALE statistic. The tibble has the following columns:
* \code{term} or \code{term1} and \code{term2}: Same as for \code{stats = 'estimate'}.
* \code{estimate}, \code{mean}, \code{median}: Same as above for individual ALE statistics.
* \code{conf.low}, \code{conf.high}: Same as above for individual ALE statistics.
}

\item{\code{what = 'ale'} (default) and \code{stats = 'conf_regions'}}{A list with elements \code{d1} and \code{d2} with the confidence regions for the requested variables and interactions. Each element is a list with the requested \code{d1} and \code{d2} sub-elements as described in the general structure above. Each data element is a tibble with confidence regions for a single variable or interaction. For an explanation of the columns, see \code{vignette('ale-statistics')}.
}

\item{\code{what = 'ale'} (default) and \code{stats = 'conf_sig'}}{A list with elements \code{d1} and \code{d2} filtered for the statistically significant confidence regions for the requested variables and interactions. Each element is a tibble. Each row is a single statistically significant confidence region. If there are no statistically significant confidence regions at all, an empty tibble is returned. For an explanation of the columns, see \code{vignette('ale-statistics')}.
}
}
}
\description{
Retrieve specific elements from an \code{ALE} object.
}
\examples{
# See examples at ALE() for a demonstration us the get() method.


}
