% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get.ALEPlots}
\alias{get.ALEPlots}
\title{get method for ALEPlots objects}
\arguments{
\item{obj}{ALEPlots object from which to retrieve ALE elements.}

\item{type}{character(1). What type of ALEPlots to retrieve: \code{'ale'} for standard ALE plots or \code{'eff'} for ALE effects plots.}
}
\value{
See \code{\link[=get.ALE]{get.ALE()}}
}
\description{
Retrieve specific plots from a \code{ALEPlots} object.

See \code{\link[=get.ALE]{get.ALE()}} for explanation of parameters not described here.
}
