% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlot_compatibility.R
\name{idxs_kolmogorov_smirnov}
\alias{idxs_kolmogorov_smirnov}
\title{Sorted categorical indices based on Kolmogorov-Smirnov distances for empirically ordering categorical categories.}
\usage{
idxs_kolmogorov_smirnov(X, x_col, n_bins, x_int_counts)
}
\arguments{
\item{X}{X data}

\item{x_col}{character}

\item{n_bins}{integer}

\item{x_int_counts}{bin sizes}
}
\description{
Sorted categorical indices based on Kolmogorov-Smirnov distances for empirically ordering categorical categories.
}
