% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBoot.R
\name{ModelBoot}
\alias{ModelBoot}
\title{A statistics and ALE data of a bootstrapped model}
\usage{
ModelBoot(
  data,
  model,
  ...,
  model_call_string = NULL,
  model_call_string_vars = character(),
  parallel = future::availableCores(logical = FALSE, omit = 1),
  model_packages = NULL,
  y_col = NULL,
  binary_true_value = TRUE,
  pred_fun = function(object, newdata, type = pred_type) {
     stats::predict(object =
    object, newdata = newdata, type = type)
 },
  pred_type = "response",
  boot_it = 100,
  seed = 0,
  boot_alpha = 0.05,
  boot_centre = "mean",
  output = c("ale", "model_stats", "model_coefs"),
  ale_options = list(),
  tidy_options = list(),
  glance_options = list(),
  silent = FALSE
)
}
\arguments{
\item{data}{dataframe. Dataset that will be bootstrapped.}

\item{model}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{...}{not used. Inserted to require explicit naming of subsequent arguments.}

\item{model_call_string}{character string. If NULL, the \code{ModelBoot} tries to automatically detect and construct the call for bootstrapped datasets. If it cannot, the function will fail early. In that case, a character string of the full call for the model must be provided that includes \code{boot_data} as the data argument for the call. See examples.}

\item{model_call_string_vars}{character. Character vector of names of variables
included in \code{model_call_string} that are not columns in \code{data}.
If any such variables exist, they must be specified here or else parallel processing
will produce an error. If parallelization is disabled with \code{parallel = 0},
then this is not a concern.}

\item{parallel}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{model_packages}{See documentation for \code{\link[=ALE]{ALE()}}}

\item{y_col, pred_fun, pred_type}{See documentation for \code{\link[=ALE]{ALE()}}. Only used to calculate bootstrapped performance measures. If NULL (default), then the relevant performance measures are calculated only if these arguments can be automatically detected.}

\item{binary_true_value}{any single atomic value. If the model represented by \code{model} or \code{model_call_string} is a binary classification model, \code{binary_true_value} specifies the value of \code{y_col} (the target outcome) that is considered \code{TRUE}; any other value of \code{y_col} is considered \code{FALSE}. This argument is ignored if the model is not a binary classification model. For example, if 2 means \code{TRUE} and 1 means \code{FALSE}, then set \code{binary_true_value} as \code{2}.}

\item{boot_it}{integer from 0 to Inf. Number of bootstrap iterations.
If boot_it = 0, then the model is run as normal once on the full \code{data} with
no bootstrapping.}

\item{seed}{integer. Random seed. Supply this between runs to assure identical
bootstrap samples are generated each time on the same data.}

\item{boot_alpha}{numeric. The confidence level for the bootstrap confidence intervals is
1 - boot_alpha. For example, the default 0.05 will give a 95\% confidence
interval, that is, from the 2.5\% to the 97.5\% percentile.}

\item{boot_centre}{See See documentation for \code{\link[=ALE]{ALE()}}}

\item{output}{character vector. Which types of bootstraps to calculate and return:
\itemize{
\item 'ale': Calculate and return bootstrapped ALE data.
\item 'model_stats': Calculate and return bootstrapped overall model statistics.
\item 'model_coefs': Calculate and return bootstrapped model coefficients.
\item 'boot_data': Return full data for all bootstrap iterations. This data will always be calculated
because it is needed for the bootstrap averages. By default, it is not returned
except if included in this \code{output} argument.
}}

\item{ale_options, tidy_options, glance_options}{list of named arguments. Arguments to pass to the \code{\link[=ALE]{ALE()}}, \code{\link[broom:reexports]{broom::tidy()}}, or \code{\link[broom:reexports]{broom::glance()}} functions, respectively, beyond (or overriding) the defaults. In particular, to obtain p-values for ALE statistics, see the details.}

\item{silent}{See documentation for \code{\link[=ALE]{ALE()}}}
}
\value{
list with the following elements (depending on values requested in the \code{output} argument:
\itemize{
\item \code{model_stats}: tibble of bootstrapped results from \code{\link[broom:reexports]{broom::glance()}}
\item \code{boot_valid}: named vector of advanced model performance measures; these are bootstrap-validated with the .632 correction (NOT the .632+ correction):
\itemize{
\item mae: mean absolute error (bootstrap validated)
\item mad: mean absolute deviation about the mean (this is a descriptive statistic calculated on the full dataset; it is provided for reference)
\item sa_mae_mad: standardized accuracy of the MAE referenced on the MAD (bootstrap validated)
\item rmse: root mean squared error (bootstrap validated)
\item standard deviation (this is a descriptive statistic calculated on the full dataset; it is provided for reference)
\item sa_rmse_sd: standardized accuracy of the RMSE referenced on the SD (bootstrap validated)
}
\item \code{model_coefs}: tibble of bootstrapped results from \code{\link[broom:reexports]{broom::tidy()}}
\item \code{ale}: list of bootstrapped ALE results
\itemize{
\item \code{data}: ALE data (see \code{\link[=ALE]{ALE()}} for details about the format)
\item \code{stats}: ALE statistics. The same data is duplicated with different views that might be variously useful:
\itemize{
\item \code{by_term}: statistic, estimate, conf.low, median, mean, conf.high. ("term" means variable name.) The column names are compatible with the \code{broom} package. The confidence intervals are based on the \code{\link[=ALE]{ALE()}} function defaults; they can be changed with the \code{ale_options} argument. The estimate is the median or the mean, depending on the \code{boot_centre} argument.
\item \code{by_stat} : term, estimate, conf.low, median, mean, conf.high.
\item \code{estimate}: term, then one column per statistic provided with the default estimate. This view does not present confidence intervals.
}
}
\item \code{boot_data}: full bootstrap data (not returned by default)
\item other values: the \code{boot_it}, \code{seed}, \code{boot_alpha}, and \code{boot_centre} arguments that were originally passed are returned for reference.
}
}
\description{
No modelling results, with or without ALE, should be considered reliable without being bootstrapped. For large datasets, normally the model provided to \code{\link[=ALE]{ALE()}} is the final deployment model that has been validated and evaluated on training and testing on subsets; that is why \code{\link[=ALE]{ALE()}} is calculated on the full dataset. However, when a dataset is too small to be subdivided into training and test sets for a standard machine learning process, then the entire model should be bootstrapped. That is, multiple models should be trained, one on each bootstrap sample. The reliable results are the average results of all the bootstrap models, however many there are. For details, see the vignette on small datasets or the details and examples below.

\code{\link[=ModelBoot]{ModelBoot()}} automatically carries out full-model bootstrapping suitable for small datasets. Specifically, it:
\itemize{
\item Creates multiple bootstrap samples (default 100; the user can specify any number);
\item Creates a model on each bootstrap sample;
\item Calculates model overall statistics, variable coefficients, and ALE values for each model on each bootstrap sample;
\item Calculates the mean, median, and lower and upper confidence intervals for each of those values across all bootstrap samples.
}
}
\section{p-values}{

The \code{\link[broom:reexports]{broom::tidy()}} summary statistics will provide p-values. However, the procedure for obtaining p-values for ALE statistics is very slow: it involves retraining the model 1000 times. Thus, it is not efficient to calculate p-values whenever a \code{ModelBoot} object is created. Although the \code{\link[=ALE]{ALE()}} function provides an 'auto' option for creating p-values, that option is disabled when creating a \code{ModelBoot} because it would be far too slow: it would involve retraining the model 1000 times the number of bootstrap iterations. Rather, you must first create a p-values distribution object using the procedure described in \code{help(-ALEpDist)}. If the name of your p-values object is \code{p_dist}, you can then request p-values each time you create a \code{ModelBoot} by passing it the argument \code{ale_options = list(p_values = p_dist)}.
}

\examples{

# attitude dataset
attitude

## ALE for general additive models (GAM)
## GAM is tweaked to work on the small dataset.
gam_attitude <- mgcv::gam(rating ~ complaints + privileges + s(learning) +
                            raises + s(critical) + advance,
                          data = attitude)
summary(gam_attitude)

\donttest{
# Full model bootstrapping
# Only 4 bootstrap iterations for a rapid example; default is 100
# Increase value of boot_it for more realistic results
mb_gam <- ModelBoot(
  attitude,
  gam_attitude,
  boot_it = 4
)

# If the model is not standard, supply model_call_string with
# 'data = boot_data' in the string (not as a direct argument to the `ModelBoot` constructor)
mb_gam <- ModelBoot(
  attitude,
  gam_attitude,
  model_call_string = 'mgcv::gam(
    rating ~ complaints + privileges + s(learning) +
      raises + s(critical) + advance,
    data = boot_data
  )',
  boot_it = 4
)

# Model statistics and coefficients
mb_gam$model_stats
mb_gam$model_coefs

# Plot ALE
mb_gam_plots <- plot(mb_gam)
mb_gam_1D_plots <- mb_gam_plots$distinct$rating$plots[[1]]
patchwork::wrap_plots(mb_gam_1D_plots, ncol = 2)
}

}
\references{
Okoli, Chitu. 2023. “Statistical Inference Using Machine Learning and Classical Techniques Based on Accumulated Local Effects (ALE).” arXiv. \url{https://arxiv.org/abs/2310.09877}.
}
