% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALEPlots_methods.R
\name{plot-ALEPlots-method}
\alias{plot-ALEPlots-method}
\title{Plot method for ALEPlots object}
\value{
Invisibly returns \code{x}.
}
\description{
Plot an \code{ALEPlots} object with the \code{\link[graphics:plot]{graphics::plot()}} generic.
}
\section{Method usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{plot(x, max_print = 20L, ...)
}\if{html}{\out{</div>}}
}

\section{Method arguments}{

\itemize{
\item \code{x}: An object of class \code{ALEPlots}.
\item \code{max_print}: integer(1). The maximum number of plots that may be printed at a time. 1D plots and 2D are printed on separate pages, so this maximum applies separately to each dimension of ALE plots, not to all dimensions combined.
\item \code{comp} See documentation for \link{get-ALE-method}
\item \code{...}: Arguments to pass to \code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}
}
}

