% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ale}
\alias{plot.ale}
\title{plot method for \code{ale} objects}
\usage{
\method{plot}{ale}(ale_obj, type = "ale")
}
\arguments{
\item{ale_obj}{ale object. The object of class \code{ale} containing data to be plotted.}

\item{character(1).}{'ale' for regular ALE plots; 'effects' for an ALE statistic effects plot.}
}
\description{
plot method for \code{ale} objects
}
