% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ale}
\alias{plot.ale}
\title{plot method for \code{ale} objects}
\usage{
\method{plot}{ale}(
  ale_obj,
  type = "ale",
  ...,
  relative_y = "median",
  p_alpha = c(0.01, 0.05),
  median_band_pct = c(0.05, 0.5),
  min_rug_per_interval = 1,
  seed = 0
)
}
\arguments{
\item{ale_obj}{ale object. The object of class \code{ale} containing data to be plotted.}

\item{character(1).}{'ale' for regular ALE plots; 'effects' for an ALE statistic effects plot.}
}
\description{
plot method for \code{ale} objects
}
