% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ale_boot}
\alias{plot.ale_boot}
\title{plot method for \code{ale_boot} objects}
\usage{
\method{plot}{ale_boot}(x, ...)
}
\arguments{
\item{x}{ale_boot object.}

\item{...}{Arguments passed to \code{\link[=plot.ale]{plot.ale()}}}
}
\description{
plot method for \code{ale_boot} objects
}
