% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ModelBoot}
\alias{plot.ModelBoot}
\title{plot method for \code{ModelBoot} objects}
\arguments{
\item{x}{ModelBoot object.}

\item{...}{Arguments passed to \code{\link[=plot.ALE]{plot.ALE()}}}
}
\description{
plot method for \code{ModelBoot} objects
}
