% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ale.R
\name{plot_ale}
\alias{plot_ale}
\title{Plot Accumulated Local Effects (ALE)}
\usage{
plot_ale(ale_data, feature_name, title_suffix = "custom")
}
\arguments{
\item{ale_data}{A dataframe containing the results from compute_ale, with columns
'feature' and 'ALE' expected.}

\item{feature_name}{A string specifying the name of the feature, used to label
the x-axis and to customize the plot title.}

\item{title_suffix}{A string to be added before the plot title, defaulting to "custom".
This can be used to distinguish plots based on different models or data subsets.}
}
\value{
A ggplot object representing the ALE plot, which can be displayed or further
        customized using ggplot2 functions.
}
\description{
This function generates a line plot of the ALE values computed for a feature. It allows
customization of the plot's title and automatically adjusts the x and y axes labels
according to the feature's name. The function uses ggplot2 for plotting.
}
\examples{
# Simulate data for example
set.seed(123)
data <- data.frame(x = runif(100, 0, 1), y = sin(2 * pi * runif(100, 0, 1)) + rnorm(100))
model <- lm(y ~ x, data = data)
ale_results_x1 <- compute_ale(model, data, "x", 10)
plot_ale(ale_results_x1, "x", "Model ALE")

}
