% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{print.ALE}
\alias{print.ALE}
\title{print Method for ALE object}
\arguments{
\item{x}{An object of class \code{ALE}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ALE object.
}
\examples{
\dontrun{
my_object <- structure(list(name = "Example", value = 42), class = "my_class")
print(my_object)
}

}
