% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ale_methods.R
\name{print.ale}
\alias{print.ale}
\title{Print Method for ale object}
\usage{
\method{print}{ale}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ale}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ale object.
}
\examples{
\dontrun{
my_object <- structure(list(name = "Example", value = 42), class = "my_class")
print(my_object)
}

}
