% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{print.ALE}
\alias{print.ALE}
\title{print Method for ALE object}
\arguments{
\item{x}{An object of class \code{ALE}.}

\item{details}{logical(1). If \code{TRUE} (default), all brief details are printed. If \code{FALSE}, only minimal information is printed.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ALE object.
}
\examples{
\donttest{
lm_cars <- stats::lm(mpg ~ ., mtcars)
ale_cars <- ALE(lm_cars, p_values = NULL)
print(ale_cars)
}

}
