% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{print.ale_plots}
\alias{print.ale_plots}
\title{Print method for ale_plots object}
\usage{
\method{print}{ale_plots}(x, max_print = 20L, ...)
}
\arguments{
\item{x}{An object of class \code{ale_plots}.}

\item{max_print}{integer(1). The maximum number of plots that may be printed at a time. 1D plots and 2D are printed separately, so this maximum applies separately to each dimension of ALE plots, not to all dimensions combined.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an ale_plots object.
}
\examples{
\dontrun{
my_object <- structure(list(name = "Example", value = 42), class = "my_class")
print(my_object)
}

}
