% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staccuracy.R
\name{mae}
\alias{mae}
\alias{rmse}
\alias{mad}
\alias{aucroc}
\title{Mean absolute error (MAE)}
\usage{
mae(actual, pred, na.rm = FALSE)

rmse(actual, pred, na.rm = FALSE)

mad(x, na.rm = FALSE, version = "mean", ...)

aucroc(
  actual,
  pred,
  na.rm = FALSE,
  binary_true_value = NULL,
  sample_size = 10000,
  seed = 0
)
}
\arguments{
\item{actual}{any atomic vector. Actual label values from a dataset. They must be binary; that is, there must be exactly two distinct values (other than missing values, which are allowed). The "true" or "positive" class is determined by coercing \code{actual} to logical \code{TRUE} and \code{FALSE} following the rules of \code{\link[=as.logical]{as.logical()}}. If this is not the intended meaning of "positive", then specify which of the two values should be considered \code{TRUE} with the argument \code{binary_true_value}.}

\item{pred}{numeric vector. Predictions corresponding to each respective element in \code{actual}. Any numeric value (not only probabilities) are permissible.}

\item{na.rm}{logical(1). \code{TRUE} if missing values should be removed; \code{FALSE} if they should be retained. If \code{TRUE}, then if any element of either \code{actual} or \code{pred} is missing, its paired element will be also removed.}

\item{x}{numeric vector. Values for which to calculate the mean absolute deviation.}

\item{version}{character(1). By default (\code{version = 'mean'}), \code{mad()} returns the mean absolute deviation (MAD) of values relative to their mean. If \code{version = 'median'}, it calls the \code{stats::mad()} function instead, the median absolute deviation relative to their median. Any other value gives an error.}

\item{...}{Arguments to pass to \code{stats::mad()} if \code{version = 'median'}. See the \code{version} argument for details.}

\item{binary_true_value}{any single atomic value. The value of \code{actual} that is considered \code{TRUE}; any other value of \code{actual} is considered \code{FALSE}. For example, if \code{2} means \code{TRUE} and \code{1} means \code{FALSE}, then set \code{binary_true_value = 2}.}

\item{sample_size}{single positive integer. To keep the computation relatively rapid, when \code{actual} and \code{pred} are longer than \code{sample_size} elements, then a random sample of \code{sample_size} of \code{actual} and \code{pred} will be selected and the ROC and AUC will be calculated on this sample. To disable random sampling for long inputs, set \code{sample_size = NA}.}

\item{seed}{numeric(1). Random seed used only if \code{length(actual) > sample_size}.}
}
\value{
MAE of \code{actual} and \code{pred}. If any value in \code{actual} or \code{pred} is
\code{NA} and \code{na.rm = FALSE}, then returns \code{NA}.

RMSE of \code{actual} and \code{pred}. If any value in \code{actual} or \code{pred} is \code{NA} and \code{na.rm = FALSE}, returns \code{NA}.

MAD of the \code{x} values. If any value of \code{x} is \code{NA} and \code{na.rm = FALSE}, returns \code{NA}.

List with the following elements:
\itemize{
\item \code{roc_opt}: tibble with optimistic ROC data. "Optimistic" means that when predictions are tied, the TRUE/positive actual values are ordered before the FALSE/negative ones.
\item \code{roc_pess}: tibble with pessimistic ROC data. "Pessimistic" means that when predictions are tied, the FALSE/negative actual values are ordered before the TRUE/positive ones. Note that this difference is not merely in the sort order: when there are ties, the way that true positives, true negatives, etc. are counted is different for optimistic and pessimistic approaches. If there are no tied predictions, then \code{roc_opt} and \code{roc_pess} are identical.
\item \code{auc_opt}: area under the ROC curve for optimistic ROC.
\item \code{auc_pess}: area under the ROC curve for pessimistic ROC.
\item \code{auc}: mean of \code{auc_opt} and \code{auc_pess}. If there are no tied predictions, then \code{auc_opt}, \code{auc_pess}, and \code{auc} are identical.
\item \code{ties}: \code{TRUE} if there are two or more tied predictions; \code{FALSE} if there are no ties.
}
}
\description{
Returns the mean absolute error (MAE) of predicted values relative to the actual values.

Returns the root mean squared error (RMSE) of predicted values relative to the actual values.

Returns the mean absolute deviation (MAD) of values relative to their mean. This is useful as a default benchmark for the mean absolute error (MAE), as the standard deviation (SD) is a default benchmark for the root mean square error (RMSE).

Returns the area under the ROC curve based on comparing the predicted scores to the actual binary values. Tied predictions are handled by calculating the optimistic AUC (positive cases sorted first, resulting in higher AUC) and the pessimistic AUC (positive cases sorted last, resulting in lower AUC) and then returning the average of the two. For the ROC, a "tie" means at least one pair of \code{pred} predictions whose value is identical yet their corresponding values of \code{actual} are different. (If the value of \code{actual} are the same for identical predictions, then these are unproblematic and are not considered "ties".)
}
\details{
NOTE: This function name overrides \code{stats::mad()} (median absolute deviation relative to their median). To maintain the functionality of \code{stats::mad()}, specify the \code{version} argument.
}
