% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset.ALEPlots}
\alias{subset.ALEPlots}
\title{subset method for ALEPlots object}
\arguments{
\item{x}{An object of class \code{ALEPlots}.}

\item{...}{not used. Inserted to require explicit naming of subsequent arguments.}

\item{include}{character in c('eff'). x_cols and exclude_cols specify precisely which variables to include or exclude in the subset. However, multivariable plots like ALE effects plot are ambiguous because they cannot be subsetted to remove some existing variables. include specifies which of such plots to retain. Possibilities are:
\itemize{
\item "eff": ALE effects plot (included, that is, not subsetted, by default).
}}
}
\value{
An \code{ALEPlots} object reduced to cover only variables and interactions specified by x_cols and exclude_cols .
}
\description{
Return an \code{ALEPlots} object reduced to cover only specified variables and interactions.

See \code{\link[=get.ALE]{get.ALE()}} for explanation of parameters not described here.
}
