% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALE_methods.R
\name{summary.ALE}
\alias{summary.ALE}
\title{summary Method for ALE object}
\usage{
## S7 method for class <ale::ALE>
summary(
  object,
  stats = c("aled", "naled"),
  all_conf = FALSE,
  round_digits = 4L,
  max_rows = 100,
  ...
)
}
\arguments{
\item{object}{An object of class \code{ALE}.}

\item{stats}{character. One or more values in c("aled", "aler_min", "aler_max", "naled", "naler_min", "naler_max"): statistics to report in detail (estimate, p-values, confidence intervals). For others not listed here, only the average (mean or median) estimates are reported. The statistics will be presented in the same order as specified.}

\item{all_conf}{logical(1). By default (\code{FALSE}), only statistically significant confidence regions are reported. If \code{TRUE}, all regions are reported as well.}

\item{round_digits}{integer(1). Numbers in tables will be rounded to \code{round_digits} decimal places.}

\item{max_rows}{natural number. Maximum number of rows to print for any component.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Invisibly returns \code{object}. The printout is a side effect.
}
\description{
Prints out a statistical summary of an \code{ALE} object. If there are no ALE statistics, a message says so. Summarized statistics are mean or median depending on the \code{boot_centre} argument used for \code{\link[=ALE]{ALE()}} bootstrapping.
}
\examples{
\donttest{
lm_cars <- stats::lm(mpg ~ ., mtcars)
ale_cars <- ALE(lm_cars, boot_it = 3)
summary(ale_cars)
}

}
