% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ALEPlots}
\alias{summary.ALEPlots}
\title{summary method for ALEPlots object}
\arguments{
\item{object}{An object of class \code{ALEPlots}.}

\item{...}{Not used}
}
\value{
Summary string.
}
\description{
Present concise summary information about an \code{ALEPlots} object.
}
